requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1

